Fit old page to updated template
