+++ src/components/pages/Settings.vue
@@ -1,19 +1,23 @@
 <template>
-  <div>
-	  <header>
-      <div class="title-card">
-        <h3>Pebble Type</h3>
-      </div>
-	  </header>
-    <main class="text-center">
-		  <form>
-			  <input type="radio" id="aplite" value="aplite" v-model="platform"> <label for="aplite">Pebble Original/Steel (aplite)</label><br />
-			  <input type="radio" id="basalt" value="basalt" v-model="platform"> <label for="basalt">Pebble Time Steel (basalt)</label><br />
-			  <input type="radio" id="chalk" value="chalk" v-model="platform"> <label for="chalk">Pebble Time Round (chalk)</label><br />
-			  <input type="radio" id="diorite" value="diorite" v-model="platform"> <label for="diorite">Pebble 2 (diorite)</label><br />
-		  </form>
-	  </main>
-  </div>
+    <div>
+        <header>
+            <div class="title-card">
+                <h3>Pebble Type</h3>
+            </div>
+        </header>
+        <main class="text-center">
+            <form>
+                <input type="radio" id="aplite" value="aplite" v-model="platform"> <label for="aplite">Pebble
+                Original/Steel (aplite)</label><br/>
+                <input type="radio" id="basalt" value="basalt" v-model="platform"> <label for="basalt">Pebble Time Steel
+                (basalt)</label><br/>
+                <input type="radio" id="chalk" value="chalk" v-model="platform"> <label for="chalk">Pebble Time Round
+                (chalk)</label><br/>
+                <input type="radio" id="diorite" value="diorite" v-model="platform"> <label for="diorite">Pebble 2
+                (diorite)</label><br/>
+            </form>
+        </main>
+    </div>
 </template>
 
 <script>
